add change pw modal, fix delete event in main page

+++ frontend/src/components/EventCard.js
@@ -4,7 +4,7 @@ import PlaceIcon from '@mui/icons-material/Place';
 import CalendarMonthIcon from '@mui/icons-material/CalendarMonth';
 import AccessTimeIcon from '@mui/icons-material/AccessTime';
 import Button from '@mui/material/Button';
-import { Paper, CardMedia } from '@mui/material'; 
+import { Paper, CardMedia } from '@mui/material';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
 import washuLogo from '../assets/washuLogo.png';
@@ -16,25 +16,21 @@ const EventCard = ({ event, onDelete }) => {
   const navigate = useNavigate();
 
   const handleSave = async () => {
-      try {
-          const response = await axios.put(`http://localhost:8000/auth/${currUser}/save-event/`, {
-              eventId: id
-          });
-          alert(response.data.message);
-      } catch (error) {
-          console.error("Error saving event:", error);
-          alert("An error occurred while saving the event.");
-      }
+    try {
+      const response = await axios.put(`http://localhost:8000/auth/${currUser}/save-event/`, {
+        eventId: id
+      });
+      alert(response.data.message);
+    } catch (error) {
+      console.error("Error saving event:", error);
+      alert("An error occurred while saving the event.");
+    }
   };
 
   const handleDelete = async () => {
-    try {
-      await axios.delete(`http://localhost:8000/events/${id}`);
-      alert('Event deleted successfully');
+    const confirmDelete = window.confirm("Are you sure you want to delete this event? This action cannot be undone.");
+    if (confirmDelete) {
       onDelete(id);
-    } catch (error) {
-      console.error("Error deleting event:", error);
-      alert("An error occurred while deleting the event.");
     }
   };
 
@@ -47,23 +43,23 @@ const EventCard = ({ event, onDelete }) => {
   };
 
   return (
-    <Paper className="event-card" onClick={handleCardClick}>
-     
-      <CardMedia
-        component="img"
-        sx={{ width: '100px', height: 'auto', margin: '10px auto' }}
-        image={washuLogo} 
-        alt="WashU Logo"
-      />
-      <h2 className="event-title"><strong>{name}</strong></h2>
-      <p className="event-description">{username}</p>
-      <p className="event-date"><CalendarMonthIcon sx={{ marginRight: 1, verticalAlign: 'middle' }} fontSize="small" /> {date}</p>
-      <p className="event-time"><AccessTimeIcon sx={{ marginRight: 1, verticalAlign: 'middle' }} fontSize="small" /> {time}</p>
-      <p className="event-location"><PlaceIcon sx={{ marginRight: 1, verticalAlign: 'middle' }} fontSize="small" /> {address}</p>
-      <p className="event-description">{details_of_event}</p>
-      
+    <Paper className="event-card">
+      <div className="event-details" onClick={handleCardClick}>
+        <CardMedia
+          component="img"
+          sx={{ width: '100px', height: 'auto', margin: '10px auto' }}
+          image={washuLogo}
+          alt="WashU Logo"
+        />
+        <h2 className="event-title"><strong>{name}</strong></h2>
+        <p className="event-description">{username}</p>
+        <p className="event-date"><CalendarMonthIcon sx={{ marginRight: 1, verticalAlign: 'middle' }} fontSize="small" /> {date}</p>
+        <p className="event-time"><AccessTimeIcon sx={{ marginRight: 1, verticalAlign: 'middle' }} fontSize="small" /> {time}</p>
+        <p className="event-location"><PlaceIcon sx={{ marginRight: 1, verticalAlign: 'middle' }} fontSize="small" /> {address}</p>
+        <p className="event-description">{details_of_event}</p>
+      </div>
       <Button size="small" color="primary" onClick={handleSave} sx={{ marginTop: 'auto' }}> Save </Button>
-      
+
       {/* Show Edit and Delete buttons only if the current user is the creator of the event */}
       {username === currUser && (
         <>
